Make aml_rdpciaddr work better through PCI bridges
When looking up the address of a PCI device from which to read for
an opregion, walk from the root inward and each time a PCI bridge is
encountered, read its bus info register and note its bus id.

Previously we were walking outward from the device and only changing
buses if there was a _BBN node which is not always the case.

This should fix booting of late Intel Macs which have AML that tries
to read the PCI class of the SSD after powering up the PCI port in
its _PS0 method, waiting in a loop for it to read 1 (Mass Storage).
Since we were previously reading the PCI class of the bridge instead
and always getting 6, this loop would have to timeout which took
many minutes.  Now we are properly reading from the device itself
which is on a different bus, and the PCI class reads correctly.

feedback from miod
ok kettenis