start moving some global pf counters to per cpu counters.
this moves the pfctl_status.fcounters, which includes the state
search counter which is bumped every time the state table is searched.

the places that get a copy of these counters now have to read the
per cpu counters and fold them into a struct pf_status rather than
just copy the global now.

ok sashan@