fix some bpf code.
jsg@ pointed out that i was unconditionally calling bpf_mtap when
the convention is to test if the bpf if pointer is set before calling
bpf_mdap.

while here drop incoming packets if the bpf filter tells us to to
be consistent with the if_input handlers.

original nit pointed out by jsg@