replace SRPs with SMRs for carp iface list hanging off an ethernet iface
this is possible now that if_vinput doesnt sleep.

carp_input is called by if_ether_input, which iterates over the
list of carp interface and checks if one of them should get the
packet, or if they should all get a copy of a multicast packet.
this traversal was done over an SRP list, but this replaces that
with an SMR based list. either way, if a carp interface is supposed
to receive a packet, it's done via a call to if_vinput during this
traversal.

i'm trying to replace SRPs with SMRs, but SMR critical sections
don't let you sleep, which if_vinput could do cos it used to run
the network stack.