netinet/ip.h: fix IPTOS_DSCP_VA value to be 0xb0
As the comment preceding this group of macros says: these are the 6 most
significant bits as they appear on the wire, so it's 0x2c << 2 == 0xb0,
matching the other entries in this list. This results in moving it to a
place that makes a bit more sense. Update comment to include RFC 5685.

Compare with RFC 2474, section 3, RFC 5685, section 4 and the values in
https://www.iana.org/assignments/dscp-registry/dscp-registry.xhtml

ok stsp